Implemented count for each item

+++ app/page.js
@@ -35,7 +35,7 @@ export default function Home() {
     const docs = await getDocs(snapshot)
     const pantryList = []
     docs.forEach((doc) => {
-      pantryList.push(doc.id)
+      pantryList.push({name:doc.id, ...doc.data()})
     })
     console.log(pantryList)
     setPantry(pantryList)
@@ -47,13 +47,28 @@ export default function Home() {
 
   const addItem = async (item) =>{
     const docRef =  doc(collection(firestore,'pantry'),item)
-    await setDoc(docRef, {})
+    const docSnap = await getDoc(docRef)
+    if(docSnap.exists()){
+      const {count} = docSnap.data()
+      await setDoc(docRef, {count:count+1})
+    }
+    else{
+      await setDoc(docRef,{count:1})
+    }
     await updatePantry()
   }
 
   const removeItem = async (item) =>{
     const docRef = doc(collection(firestore,'pantry'), item)
-    await deleteDoc(docRef)
+    const docSnap = await getDoc(docRef)
+    if(docSnap.exists()){
+      const {count} = docSnap.data()
+      if(count==1){
+        await deleteDoc(docRef)
+      } else{
+        await setDoc(docRef, {count:count-1})
+      }
+    }
     await updatePantry()
   }
 
@@ -86,7 +101,7 @@ export default function Home() {
             onChange={(e) => setItemName(e.target.value)}/>
             <Button variant="outlined" 
             onClick={()=>{
-              addItem(itemName)
+              addItem(itemName.charAt(0).toUpperCase() + itemName.slice(1))
               setItemName('')
               handleClose()
             }}>
@@ -109,9 +124,9 @@ export default function Home() {
           </Typography>
         </Box>
         <Stack width="800px"height="300px"spacing={2} overflow={'auto'}>
-        {pantry.map((i)=>(
+        {pantry.map(({name, count})=>(
           <Box
-            key={i}
+            key={name}
             width="100%"
             minHeight="150px"
             display={'flex'}
@@ -120,14 +135,13 @@ export default function Home() {
             bgcolor={'#F0F0F0'}
             paddingX={5}
           >
-            <Typography
-              variant={'h3'}
-              color={'#333'}
-              textAlign={'center'}
-            >
-              {i.charAt(0).toUpperCase() + i.slice(1)}
+            <Typography variant={'h3'} color={'#333'} textAlign={'center'}>
+              {name.charAt(0).toUpperCase() + name.slice(1)}
             </Typography>
-            <Button variant='contained' onClick={() => removeItem(i)}>Remove</Button>
+            <Typography variant={'h3'} color={'#333'} textAlign={'center'}>
+              Quantity: {count}
+            </Typography>
+            <Button variant='contained' onClick={() => removeItem(name)}>Remove</Button>
           </Box>
         ))}
         </Stack>
